création du dao abstract et du generic services

+++ src/main/java/fr/dta/mediatic/adherent/dao/AdherentDao.java
@@ -39,7 +39,7 @@ public class AdherentDao extends AbstractDao<Adherent> {
 		return super.getAll(table);
 	}
 	
-	public Adherent getById(int id){
+	public Adherent getById(Long id){
 		return super.getById(id, table);
 	}
